other_langs/tests_cdk/s3_eventbridge/requirements.txt
aws-cdk-lib==2.89.0
constructs>=10.0.0,<11.0.0